+++ google/cloud/testing_util/CMakeLists.txt
@@ -65,7 +65,7 @@ if (BUILD_TESTING OR GOOGLE_CLOUD_CPP_TESTING_UTIL_ENABLE_INSTALL)
 
     find_package(googleapis)
     add_library(google_cloud_cpp_testing_grpc is_proto_equal.cc
-                                              is_proto_equal.h)
+                                              is_proto_equal.h mock_completion_queue.h mock_async_response_reader.h)
     target_link_libraries(
         google_cloud_cpp_testing_grpc
         PUBLIC google_cloud_cpp_common protobuf::libprotobuf GTest::gmock
